ci: split agents-host deployment tests across shards
